+++ NEWS.md
@@ -1,3 +1,10 @@
+# LRTesteR 1.5.0
+* gaussian_mu_one_way and log_normal_mu_one_way no longer assume equal variance across groups.
+
+# LRTesteR 1.4.0
+* Add empirical likelihood tests for variance
+* Redo all vignettes
+
 # LRTesteR 1.3.1
 * Fix greek letter in readme for better display on cran.